Remove sbuf based rep, osql, newsql, and fastsql.
