+++ db/comdb2.h
@@ -1917,7 +1917,6 @@ int thd_init(void);
 void thd_cleanup();
 void sqlinit(void);
 void sqlnet_init(void);
-int clnt_stats_init(void);
 int sqlpool_init(void);
 int schema_init(void);
 int osqlpfthdpool_init(void);
