+++ db/db_tunables.c
@@ -208,7 +208,6 @@ extern int gbl_processor_thd_poll;
 extern int gbl_time_rep_apply;
 extern int gbl_incoherent_logput_window;
 extern int gbl_dump_net_queue_on_partial_write;
-extern int gbl_dump_full_net_queue;
 extern int gbl_debug_partial_write;
 extern int gbl_debug_sleep_on_verify;
 extern int gbl_max_clientstats_cache;
@@ -281,7 +280,6 @@ extern int gbl_instrument_dblist;
 extern int gbl_replicated_truncate_timeout;
 extern int gbl_match_on_ckp;
 extern int gbl_verbose_set_sc_in_progress;
-extern int gbl_send_failed_dispatch_message;
 extern int gbl_logdelete_lock_trace;
 extern int gbl_flush_log_at_checkpoint;
 extern int gbl_online_recovery;
@@ -444,7 +442,6 @@ extern int gbl_debug_sleep_in_sql_tick;
 extern int gbl_debug_sleep_in_analyze;
 extern int gbl_debug_sleep_in_summarize;
 extern int gbl_debug_sleep_in_trigger_info;
-extern int gbl_protobuf_prealloc_buffer_size;
 extern int gbl_replicant_retry_on_not_durable;
 extern int gbl_enable_internal_sql_stmt_caching;
 extern int gbl_longreq_log_freq_sec;
