--- tests/basic.test/nolibevent.testopts
@@ -1,1 +1,0 @@
-libevent off
