--- tests/electbug.test/lrl.options
@@ -1,2 +1,0 @@
-libevent 1
-libevent_appsock 1
