--- tests/mem_protobuf.test/lrl.options
@@ -1,1 +1,0 @@
-libevent_appsock 1
