--- tests/mem_protobuf.test/sbuf.testopts
@@ -1,1 +1,0 @@
-libevent_appsock 0
