+++ tests/protobuf_to_fd.test/lrl.options
@@ -1,3 +1,2 @@
 sqlenginepool mint 0
 sqlenginepool maxt 1
-libevent_appsock 1
