--- tests/ssl_rep.test/sbuf2.testopts
@@ -1,3 +1,0 @@
-libevent off
-ssl_client_mode VERIFY_CA
-ssl_replicant_mode VERIFY_CA
